Modify fmail send mail

+++ Core/Controller/SendMail.php
@@ -92,12 +92,17 @@ class SendMail extends Controller
      */
     public function url()
     {
-        $fileName = $this->request->get('fileName', '');
-        if (empty($fileName)) {
+        $sendParams['fileName'] = $this->request->get('fileName', '');
+        if (empty($sendParams['fileName'])) {
             return parent::url();
         }
 
-        return parent::url() . '?fileName=' . $fileName;
+        if ($this->request->get('modelClassName') && $this->request->get('modelCode')) {
+            $sendParams['modelClassName'] = $this->request->get('modelClassName');
+            $sendParams['modelCode'] = $this->request->get('modelCode');
+        }
+
+        return parent::url() . '?' . http_build_query($sendParams);
     }
 
     /**
@@ -164,10 +169,35 @@ class SendMail extends Controller
         }
        
         if ($emailTools->send($mail)) {
-            unlink(FS_FOLDER . '/MyFiles/' . $fileName);
+            if (\file_exists(FS_FOLDER . '/MyFiles/' . $fileName)) {
+                unlink(FS_FOLDER . '/MyFiles/' . $fileName);
+            }
+            $this->updateFemail();
             $this->miniLog->info('send-mail-ok');
         } else {
             $this->miniLog->error('send-mail-error');
+        }
+    }
+
+    /**
+     * Update the property femail with actual date if exist param ModelClassName and ModelCode
+     *
+     * @return void
+     */
+    private function updateFemail() : void
+    {
+        $className = '\FacturaScripts\Core\Model\\' . $this->request->get('modelClassName');
+        if(class_exists($className)) {
+            $model = new $className();
+            $modelCode = $this->request->get('modelCode');
+            if($model->loadFromCode($modelCode)) {
+                if (property_exists($className, 'femail') ) {
+                    $model->femail = date('d-m-Y');
+                    if (!$model->save()) {
+                        $this->miniLog->alert('error-saving-data');
+                    }
+                }
+            }
         }
     }
